chore(connect): reorganize types
