+++ packages/connect/src/types/api/cardanoGetNativeScriptHash.ts
@@ -1,25 +1,5 @@
-import type { PROTO } from '../../constants';
 import type { Params, Response } from '../params';
-
-export interface CardanoNativeScript {
-    type: PROTO.CardanoNativeScriptType;
-    scripts?: CardanoNativeScript[];
-    keyHash?: string;
-    keyPath?: string | number[];
-    requiredSignaturesCount?: number;
-    invalidBefore?: string;
-    invalidHereafter?: string;
-}
-
-export interface CardanoGetNativeScriptHash {
-    script: CardanoNativeScript;
-    displayFormat: PROTO.CardanoNativeScriptHashDisplayFormat;
-    derivationType?: PROTO.CardanoDerivationType;
-}
-
-export interface CardanoNativeScriptHash {
-    scriptHash: string;
-}
+import type { CardanoGetNativeScriptHash, CardanoNativeScriptHash } from './cardano';
 
 export declare function cardanoGetNativeScriptHash(
     params: Params<CardanoGetNativeScriptHash>,
