+++ packages/connect/src/types/api/cardanoGetPublicKey.ts
@@ -1,18 +1,5 @@
-import type { PROTO } from '../../constants';
 import type { Params, BundledParams, Response } from '../params';
-
-export interface CardanoGetPublicKey {
-    path: string | number[];
-    showOnTrezor?: boolean;
-    derivationType?: PROTO.CardanoDerivationType;
-}
-
-export interface CardanoPublicKey {
-    path: number[];
-    serializedPath: string;
-    publicKey: string;
-    node: PROTO.HDNodeType;
-}
+import type { CardanoGetPublicKey, CardanoPublicKey } from './cardano';
 
 export declare function cardanoGetPublicKey(
     params: Params<CardanoGetPublicKey>,
