+++ packages/connect/src/types/api/eosGetPublicKey.ts
@@ -1,16 +1,5 @@
 import type { Params, BundledParams, Response } from '../params';
-
-export interface EosGetPublicKey {
-    path: string | number[];
-    showOnTrezor?: boolean;
-}
-
-export interface EosPublicKey {
-    wifPublicKey: string;
-    rawPublicKey: string;
-    path: number[];
-    serializedPath: string;
-}
+import { EosGetPublicKey, EosPublicKey } from './eos';
 
 export declare function eosGetPublicKey(params: Params<EosGetPublicKey>): Response<EosPublicKey>;
 export declare function eosGetPublicKey(
