+++ packages/connect/src/types/api/ethereumSignMessage.ts
@@ -1,11 +1,6 @@
 import type { PROTO } from '../../constants';
 import type { Params, Response } from '../params';
-
-export interface EthereumSignMessage {
-    path: string | number[];
-    message: string;
-    hex?: boolean;
-}
+import type { EthereumSignMessage } from './ethereum';
 
 export declare function ethereumSignMessage(
     params: Params<EthereumSignMessage>,
