+++ packages/connect/src/types/api/tezosSignTransaction.ts
@@ -1,70 +1,6 @@
 import type { PROTO } from '../../constants';
 import type { Params, Response } from '../params';
-
-export interface TezosRevealOperation {
-    source: string;
-    fee: number;
-    counter: number;
-    gas_limit: number;
-    storage_limit: number;
-    public_key: string;
-}
-
-export interface TezosManagerTransfer {
-    destination: string;
-    amount: number;
-}
-
-export interface TezosParametersManager {
-    set_delegate?: string;
-    cancel_delegate?: boolean;
-    transfer?: TezosManagerTransfer;
-}
-
-export interface TezosTransactionOperation {
-    source: string;
-    destination: string;
-    amount: number;
-    counter: number;
-    fee: number;
-    gas_limit: number;
-    storage_limit: number;
-    parameters?: number[];
-    parameters_manager?: TezosParametersManager;
-}
-
-export interface TezosOriginationOperation {
-    source: string;
-    balance: number;
-    delegate?: string;
-    script: string | number[];
-    fee: number;
-    counter: number;
-    gas_limit: number;
-    storage_limit: number;
-}
-
-export interface TezosDelegationOperation {
-    source: string;
-    delegate: string;
-    fee: number;
-    counter: number;
-    gas_limit: number;
-    storage_limit: number;
-}
-
-export interface TezosOperation {
-    reveal?: TezosRevealOperation;
-    transaction?: TezosTransactionOperation;
-    origination?: TezosOriginationOperation;
-    delegation?: TezosDelegationOperation;
-}
-
-export interface TezosSignTransaction {
-    path: string | number[];
-    branch: string;
-    operation: TezosOperation;
-}
+import type { TezosSignTransaction } from './tezos';
 
 export declare function tezosSignTransaction(
     params: Params<TezosSignTransaction>,
